requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.34,<1.0
httpx>=0.28,<1.0
google-genai>=1.0,<2.0
google-auth>=2.40,<3.0
google-auth-oauthlib>=1.2,<2.0
google-api-python-client>=2.150,<3.0
python-dotenv>=1.0,<2.0
supabase>=2.9,<3.0
tzdata>=2024.2
